+++ src/com/euhedral/game/Entities/Enemy/EnemyBasic2.java
@@ -44,8 +44,8 @@ public class EnemyBasic2 extends Enemy{
     }
 
     @Override
-    public int getTurretX() {
-        return (int) pos.x + width/2 - Utility.intAtWidth640(2);
+    public double getTurretX() {
+        return  pos.x + width/2 - Utility.intAtWidth640(2);
     }
 
     @Override
