+++ src/com/euhedral/game/Entities/Enemy/EnemyBasic3.java
@@ -70,8 +70,8 @@ public class EnemyBasic3 extends Enemy{
     }
 
     @Override
-    public int getTurretX() {
-        return (int) pos.x + width/2 - Utility.intAtWidth640(2);
+    public double getTurretX() {
+        return pos.x + width/2 - Utility.intAtWidth640(2);
     }
 
     @Override
