+++ src/com/euhedral/game/Entities/Enemy/EnemyDrone4.java
@@ -69,8 +69,8 @@ public class EnemyDrone4 extends EnemyDrone1 {
     }
 
     @Override
-    public int getTurretX() {
-        return (int) pos.x + width/2 - Utility.intAtWidth640(2);
+    public double getTurretX() {
+        return pos.x + width/2 - Utility.intAtWidth640(2);
     }
 
     @Override
