+++ src/com/euhedral/game/Entities/Enemy/EnemyMinefield1.java
@@ -113,8 +113,8 @@ public class EnemyMinefield1 extends Enemy{
     }
 
     @Override
-    public int getTurretX() {
-        return (int) pos.x + width/2 - Utility.intAtWidth640(2);
+    public double getTurretX() {
+        return pos.x + width/2 - Utility.intAtWidth640(2);
     }
 
     private void renderPath(Graphics g) {
