+++ src/com/euhedral/game/Entities/Enemy/EnemyMinefield2.java
@@ -83,8 +83,8 @@ public class EnemyMinefield2 extends Enemy{
     }
 
     @Override
-    public int getTurretY() {
-        return (int) pos.y + Utility.intAtWidth640(3);
+    public double getTurretY() {
+        return  pos.y + Utility.intAtWidth640(3);
     }
 
     @Override
