+++ src/com/euhedral/game/Entities/Enemy/EnemyScatter1.java
@@ -127,8 +127,8 @@ public class EnemyScatter1 extends Enemy {
     }
 
     @Override
-    public int getTurretX() {
-        return (int) pos.x + width / 2 - Utility.intAtWidth640(2);
+    public double getTurretX() {
+        return pos.x + width / 2 - Utility.intAtWidth640(2);
     }
 
     @Override
