+++ src/com/euhedral/game/Entities/Enemy/EnemySide1.java
@@ -82,8 +82,8 @@ public class EnemySide1 extends Enemy{
     }
 
     @Override
-    public int getTurretY() {
-        return (int) pos.y + Utility.intAtWidth640(3);
+    public double getTurretY() {
+        return pos.y + Utility.intAtWidth640(3);
     }
 
     @Override
