+++ src/com/euhedral/game/Entities/Enemy/EnemySide3.java
@@ -119,8 +119,8 @@ public class EnemySide3 extends Enemy{
     }
 
     @Override
-    public int getTurretY() {
-        return (int) pos.y + Utility.intAtWidth640(3);
+    public double getTurretY() {
+        return pos.y + Utility.intAtWidth640(3);
     }
 
     @Override
