+++ src/com/euhedral/game/Entities/Enemy/EnemyStatic1.java
@@ -63,8 +63,8 @@ public class EnemyStatic1 extends Enemy {
     }
 
     @Override
-    public int getTurretX() {
-        return (int) pos.x + width / 2 - Utility.intAtWidth640(2);
+    public double getTurretX() {
+        return pos.x + width / 2 - Utility.intAtWidth640(2);
     }
 
     @Override
@@ -103,7 +103,7 @@ public class EnemyStatic1 extends Enemy {
     }
 
     @Override
-    public int getTurretY() {
+    public double getTurretY() {
         return (int) pos.y + 12;
     }
